+++ src/TransitEx.cpp
@@ -42,13 +42,14 @@ struct TransitExModule : TransitBase<NUM_PRESETS> {
 			BASE::slot[i].presetButton = &BASE::presetButton[i];
 		}
 
-		BASE::onReset();
+		onReset();
 	}
 
 	void onReset() override { 
 		for (int i = 0; i < NUM_PRESETS; i++) {
 			BASE::presetSlotUsed[i] = false;
 			BASE::textLabel[i] = "";
+			BASE::fadeTime[i] = -1.f;
 			BASE::preset[i].clear();
 			BASE::lights[LIGHT_PRESET + (i * 3) + 0].setBrightness(0.f);
 			BASE::lights[LIGHT_PRESET + (i * 3) + 1].setBrightness(0.f);